my diary list コンポーネント化
