--- React/unknwon-react-diary/src/component/About.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import Container from '@material-ui/core/Container';
-
-const Lp = () => {
-  return (
-    <Container>
-      <p>fd</p>
-      lpです
-    </Container>
-  );
-};
-
-export default Lp;
