+++ React/unknwon-react-diary/src/component/molecules/StatisticalDataCard.js
